+++ src/lpi2c.rs
@@ -33,9 +33,23 @@ pub trait Pin: super::Iomuxc {
 ///
 /// If you do not call `prepare()` on your I2C pin, it might not work as a I2C
 /// pin.
+///
+/// In most build configurations, `prepare()` emits code to enable an open drain.
+/// This code is correct for all 10xx MCUs.
+///
+/// If *only* the 1170 MCU feature is enabled, then `prepare()` *does not* enable
+/// an open drain. For more information on this limitation, see
+/// [the issue tracker](https://github.com/imxrt-rs/imxrt-iomuxc/issues/28).
+///
+/// If multiple MCU features are enabled, then the `prepare()` assumes that the
+/// target MCU is a 10xx MCU. This code may execute on an 1170 MCU, but it may not
+/// work as expected.
+///
+/// `prepare()` will not touch any other bits in the PAD_CTL register.
 pub fn prepare<P: Pin>(pin: &mut P) {
     super::alternate(pin, P::ALT);
     super::set_sion(pin);
+    crate::config::set_open_drain(pin);
     unsafe { P::DAISY.write() };
 }