+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.2.6] 2024-01-27
+
 - Enable an open drain when preparing I2C pins on 10xx MCUs.
 
 ## [0.2.5] 2023-12-23
@@ -182,6 +184,7 @@ First release
   - UART
 
 [Unreleased]: https://github.com/imxrt-rs/imxrt-iomuxc/compare/v0.2.0...HEAD
+[0.2.6]: https://github.com/imxrt-rs/imxrt-iomuxc/compare/v0.2.5...v0.2.6
 [0.2.5]: https://github.com/imxrt-rs/imxrt-iomuxc/compare/v0.2.4...v0.2.5
 [0.2.4]: https://github.com/imxrt-rs/imxrt-iomuxc/compare/v0.2.3...v0.2.4
 [0.2.3]: https://github.com/imxrt-rs/imxrt-iomuxc/compare/v0.2.2...v0.2.3